pyproject.toml
[project]
name = "open-agent-traces"
version = "1.1.0"
description = "Agento observability with OpenTelemetry"
requires-python = ">=3.8"
dependencies = [
    "opentelemetry-sdk~=1.34",
    "opentelemetry-exporter-otlp~=1.34",
    "opentelemetry-instrumentation~=1.34",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"